refactor(87a): split core/database.py into core/database/ package with permanent facade
Mechanical split of the 2,152-line monolith into a core/database/ package
(spec-87 D0): connection / video / migrate / alias / tag_alias / actress
submodules + a permanent re-export facade __init__.py. Public import path
core.database is unchanged; all 13 names (incl. _migrate_old_aliases) re-exported
with __all__. Full suite 4815 passed, 2 skipped — identical to pre-split baseline.

All class/function bodies, docstrings, and SQL strings moved byte-exact
(machine-verified vs HEAD:core/database.py after reversing the rewrites below).

Intentional changes beyond pure move:
- CD-87a-9: sibling submodules call connection helpers via `from . import
  connection` + `connection.get_db_path()/get_connection()/init_db()` so the
  facade-level monkeypatch can reach repo-internal calls (copied-name import
  would not). logger=get_logger(__name__) only in connection.py + video.py.
- Test patch-target updates (mechanical, behavior-preserving): 6 files change
  `core.database.get_db_path` -> `core.database.connection.get_db_path`.
  Plan-87a risk #6 enumerated 4; test_local_status_api.py (7 real failures) and
  test_showcase_api.py needed the same one-string fix (same root cause) — added.
- Dropped the orphaned section-banner comment (old L1629-1631,
  "# TagAliasRecord dataclass + TagAliasRepository") that separated two classes
  now living in separate files; tag_alias.py's module docstring already conveys it.
